refactor(F0Card)!: move F0CardHorizontal to experimental
In this Storybook the "Experimental" section, the maturity alert, and the
import banner are all driven by the directory a story file lives in
(src/experimental/, via titlePrefix) — never by the `experimental` tag (its
tagBadges config is inert: the consuming addon isn't installed). So filing
F0CardHorizontal as experimental requires physically moving it there.

Move the component, its horizontal-only CardHorizontalActions, story, mdx and
test from src/components/F0Card to src/experimental/F0CardHorizontal. Shared
card parts (CardAvatar/CardAlert/CardActions/types) stay with the vertical
F0Card and are imported via @/components/F0Card. Export from the experimental
entry (@factorialco/f0-react/experimental) instead of the main entry, and
repoint the F0HILActionConfirmation consumer.

Storybook now files it under Experimental/CardHorizontal (storyId
experimental-cardhorizontal-*, which fires the "Experimental component" alert).
Chromatic re-baselines: story IDs card-horizontal-* -> experimental-cardhorizontal-*.
