+++ packages/react/src/components/F0Card/index.tsx
@@ -1,2 +1,1 @@
 export * from "./F0Card"
-export * from "./F0CardHorizontal"
