+++ packages/react/src/experimental/F0CardHorizontal/__stories__/F0CardHorizontal.stories.tsx
@@ -3,9 +3,9 @@ import type { Meta, StoryObj } from "@storybook/react-vite"
 import image from "@storybook-static/avatars/person04.jpg"
 import { fn } from "storybook/test"
 
+import { F0Card } from "@/components/F0Card"
 import { Briefcase, Check, Cross, Delete, Envelope } from "@/icons/app"
 
-import { F0Card } from "../F0Card"
 import { F0CardHorizontal } from "../F0CardHorizontal"
 
 // Story handlers alert which control fired (on top of the Actions-panel spy),
